+++ CHANGELOG.md
@@ -29,6 +29,9 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   `InstanceStatus` as an additional argument.
   `start_adding_service` has been renamed to `initiate_adding_service` to
   better distinguish between starting and stopping a service. (#1605)
+- `after_transactions` hook is now invoked on the genesis block for the builtin
+  services. Note that calling `blockchain::Schema::height` within `after_transactions`
+  hook will cause a panic for a builtin service. (#1619)
 
 - `proposer_id` field in `Block` has been moved to additional block headers. (#1602)
 
@@ -52,6 +55,10 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   Service data becomes unavailable to other services, but still exists. The name
   and identifier remain reserved for the stopped service and cannot be used again
   for adding new services. (#1605)
+- New `blockchain::Schema` method `next_height` was added as a non-panicking
+  alternative to `height`. (#1619)
+- New method `in_genesis_block` was added to the `CallContext` to check if the service
+  hook is being executed for the genesis block. (#1619)
 
 - New `api::Error` variants were added: `Gone` and `MovedPermanently`. (#1607)
 
@@ -65,6 +72,7 @@ The project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html)
   aggregation works. (#1553)
 - Added hashed version of `Entry` called `ProofEntry`, which participates
   in the state aggregation. (#1553)
+- Added support of unsized keys to `MapIndex`. (#1621)
 
 - Added mechanism to extend block header. Block now contains
   key-value storage `additional_headers` which can contain binary data. (#1602)